Extract OAuth logic to OAuthClient class.
